feat(config): allow using function as `config.git.url_format` value
This commit add support for setting the value of `config.git.url_format`
to a function. This is usefull in cases where multiple plugin
repositories are used.

For example, the following allows user to prefix a plugin name with
"github:", "gitlab:", or "codeberg:" to embed the repository
information. Non-prefixed plugin name will resolves to the default
repository (github):

```lua
{
  git = {
    url_format = function(name)
      local prefix, url = name:match("^(.*):(.*)")

      if prefix ~= nil then
         if prefix == "github" or prefix == "gitlab" then
            return string.format(" https://%s.com/%s.git", prefix, url)
         elseif prefix == "codeberg" then
            return string.format("https://codeberg.org/%s.git", url)
         end
      end

      return string.format("https://github.com/%s.git", name)
    end
  }
}
```

Signed-off-by: Jakka Prihatna <jprihatna@gmail.com>