feat(plugin): resolve self url and branch info
Prior to this commit, the only information provided for
the plugin-self are:

```lua
{
  "folke/lazy.nvim",
  dir = "install_dir",
  ...
}
```

Which, on default circumstances resolves to:

```lua
{
  "folke/lazy.nvim",
  dir = "install_dir",
  url = "https://github.com/folke/lazy.nvim.git",
  branch = "main",
  ...
}
```

This causes issue when the plugin is a fork of
the original "folke/lazy.nvim", since the expected
url and or branch are usually different.

This commit updates the plugin to dynamically
resolve its own `url` and `branch` information based
on the (install) `dir` information.

Signed-off-by: Jakka Prihatna <jprihatna@gmail.com>